Added a boolean to setElementData to allow syncing it to the clients. Bandwidth savings for SET_ELEMENT_DATA and REMOVE_ELEMENT_DATA RPCs. Bandwidth savings for lua arguments sync. Added lots of missing casts.

+++ trunk/MTA10/mods/deathmatch/logic/CClientGame.cpp
@@ -1594,7 +1594,7 @@ void CClientGame::UpdatePlayerWeapons ( void )
             unsigned short usAmmoInClip = 0;
             CWeapon* pWeapon = m_pLocalPlayer->GetWeapon ( m_lastWeaponSlot );
             if ( pWeapon )
-                usAmmoInClip = pWeapon->GetAmmoInClip ();
+                usAmmoInClip = static_cast < unsigned short > ( pWeapon->GetAmmoInClip () );
 
             // Force it back to the old slot
             m_pLocalPlayer->SetCurrentWeaponSlot ( m_lastWeaponSlot );
@@ -1635,8 +1635,8 @@ void CClientGame::UpdatePlayerWeapons ( void )
                 if ( CWeaponNames::DoesSlotHaveAmmo ( uiSlot ) )
                 {
                     SWeaponAmmoSync ammo ( pWeapon->GetType (), true, true );
-                    ammo.data.usAmmoInClip = pWeapon->GetAmmoInClip ();
-                    ammo.data.usTotalAmmo = pWeapon->GetAmmoTotal ();
+                    ammo.data.usAmmoInClip = static_cast < unsigned short > ( pWeapon->GetAmmoInClip () );
+                    ammo.data.usTotalAmmo = static_cast < unsigned short > ( pWeapon->GetAmmoTotal () );
                     BitStream.Write ( &ammo );
                 }
             }
